refactor(app): settings tabs
