--- src/module/settings/screen/SettingsScreen.styles.ts
@@ -1,7 +1,0 @@
-import { Col } from "@peersyst/react-native-components";
-import styled from "@peersyst/react-native-styled";
-
-export const SettingsScreenContentRoot = styled(Col)(() => ({
-    marginTop: 20,
-    gap: 12,
-}));
